pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "switools"
version = "1.4"
description = "Tools for handling Arista SWI/X"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jsonschema>=4.24.0",
    "m2crypto>=0.45.1",
    "pyparsing>=3.2.3",
    "pyyaml>=6.0.2",
    "typer>=0.16.0",
]

[project.scripts]
swi-tools = "switools.__main__:main"
verify-swi = "switools.verify:app"
swi-signature = "switools.signature:app"
swix-create = "switools.create:app"
swix-signature = "switools.signature:app"
verify-swix = "switools.verify:app"
swi-signing-service = "switools.signingservice:main"

[tool.hatch.build]
packages = [
    "src/switools",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
]
